event_management/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
python-decouple==3.7
django-phonenumber-field[phonenumberslite]==7.1.0
phonenumbers==8.13.13
coverage==7.0.4